BVHViewerWindowの構造化と機能分割
- BVHViewerWindow: メインウィンドウの統合・調整機能に集約
- BVHTreeView: TreeView表示とBVH階層構築を分離
- BVHDetailPanel: 詳細情報表示パネルを独立化
- BVHDebugger: デバッグ機能を静的クラスとして分離
- 各コンポーネントの単一責任化により保守性・再利用性が向上
- UIスタイル改善（交互背景色、パディング調整）
