+++ Assets/BVH/Editor/BVHDetailPanel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2b3c4d5e6f7a8b9c0d1e2f3a4b5c6d7e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
